requirements/requirements-icon-core-cuda118.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.50.0
Deprecated==1.2.18
distlib==0.3.9
docker==7.1.0
docker-pycreds==0.4.0
einops==0.8.1
fastapi==0.115.12
filelock==3.13.1
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
identify==2.6.9
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lightning==2.5.1
lightning-utilities==0.14.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mlflow==2.21.3
mlflow-skinny==2.21.3
mpmath==1.3.0
multidict==6.4.3
networkx==3.3
neuraloperator==1.0.2
nodeenv==1.9.1
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.7
pre_commit==4.2.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rootutils==1.0.7
rsa==4.9.1
ruff==0.11.6
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.26.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.6.0+cu118
torchmetrics==1.7.1
torchvision==0.21.0+cu118
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.30.0
wandb==0.19.9
Werkzeug==3.1.3
wrapt==1.17.2
yamllint==1.37.0
yarl==1.20.0
zipp==3.21.0
